Updates based on review comments
