+++ dev/src/main/java/org/eclipse/codewind/intellij/ui/module/CodewindModuleType.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2019 IBM Corporation and others.
+ * Copyright (c) 2019, 2020 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v2.0
  * which accompanies this distribution, and is available at
@@ -9,7 +9,7 @@
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
 
-package org.eclipse.codewind.intellij.module;
+package org.eclipse.codewind.intellij.ui.module;
 
 import com.intellij.openapi.module.ModuleType;
 import org.eclipse.codewind.intellij.ui.IconCache;
@@ -18,12 +18,20 @@ import org.jetbrains.annotations.NotNull;
 
 import javax.swing.*;
 
+import static org.eclipse.codewind.intellij.ui.messages.CodewindUIBundle.message;
+
 public class CodewindModuleType extends ModuleType<CodewindModuleBuilder> {
 
-    static final CodewindModuleType CODEWIND_MODULE = new CodewindModuleType();
+    public static final String CODEWIND_MODULE= "CODEWIND_MODULE";
+
+    private static final CodewindModuleType CODEWIND_MODULE_TYPE = new CodewindModuleType();
 
     private CodewindModuleType() {
-        super("CODEWIND_MODULE");
+        super(CODEWIND_MODULE);
+    }
+
+    public static CodewindModuleType getInstance() {
+        return CODEWIND_MODULE_TYPE;
     }
 
     @NotNull
@@ -36,14 +44,14 @@ public class CodewindModuleType extends ModuleType<CodewindModuleBuilder> {
     @NotNull
     @Override
     public String getName() {
-        return "Codewind Module";
+        return message("CodewindModuleName");
     }
 
     @Nls(capitalization = Nls.Capitalization.Sentence)
     @NotNull
     @Override
     public String getDescription() {
-        return "Codewind module";
+        return message("CodewindModuleDescription");
     }
 
     @Override
