+++ dev/src/main/resources/messages/ui/CodewindUIBundle.properties
@@ -359,10 +359,13 @@ RepoUpdateErrorTitle=Template Source Update Error
 RepoListErrorTitle=Template Source List Error
 RepoListErrorMsg=An error occurred while trying to get the list of template sources: {0}
 
+CodewindModuleName=Codewind Module
+CodewindModuleDescription=Codewind modules are used to develop Java applications which run in containers.
+
 CodewindNotStarted=Codewind must be running in order to create projects or modules.
 StartCodewind=Start Codewind
-CodewindStarted=Codewind has started. Press Next to continue.
+CodewindStarted=Codewind has started. To continue, click Next.
 
 CodewindNotInstalled=Codewind images must be installed in order to create projects or modules.
 InstallCodewind=Install Codewind Images
-CodewindInstalledStarted=Codewind been installed and has started. Press Next to continue.
+CodewindInstalledStarted=Codewind has been installed and has started. To continue, click Next.
